requirements.txt
Flask==2.2.5
line-bot-sdk==2.4.1
gunicorn==20.1.0
